+++ cores/fbneo/src/burn/drv/megadrive/d_megadrive.cpp
@@ -52749,4 +52749,3 @@ struct BurnDriverD BurnDrvmd_insnspeed = {
 };
 
 */
-